+++ LowerTriangularArrays/Project.toml
@@ -20,5 +20,11 @@ LinearAlgebra = "1.10"
 SpeedyWeatherInternals = "0.1"
 julia = "1.10"
 
+[extensions]
+LowerTriangularArraysFiniteDifferencesExt = "FiniteDifferences"
+
 [sources.SpeedyWeatherInternals]
 path = "../SpeedyWeatherInternals"
+
+[weakdeps]
+FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
