+++ RingGrids/Project.toml
@@ -26,5 +26,11 @@ SpeedyWeatherInternals = "0.1"
 Statistics = "1.10"
 julia = "1.10"
 
+[extensions]
+RingGridsFiniteDifferencesExt = "FiniteDifferences"
+
 [sources.SpeedyWeatherInternals]
 path = "../SpeedyWeatherInternals"
+
+[weakdeps]
+FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
