+++ SpeedyTransforms/Project.toml
@@ -46,3 +46,11 @@ RingGrids = "0.1"
 Statistics = "1.10"
 SpeedyWeatherInternals = "0.1"
 julia = "1.10"
+
+[extensions]
+SpeedyTransformsCUDAExt = "CUDA"
+SpeedyTransformsEnzymeExt = "Enzyme"
+
+[weakdeps]
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
